Move serde utils to plugin::utils::serde
