+++ apollo-router-core/src/plugin/utils/mod.rs
@@ -1,3 +1,4 @@
 //! Utilities which make it easy to work with [`crate::plugin`].
 
+pub mod serde;
 pub mod test;
