+++ apollo-router-core/src/plugins/mod.rs
@@ -5,5 +5,4 @@
 mod forbid_mutations;
 mod headers;
 mod include_subgraph_errors;
-pub mod serde_utils;
 pub(crate) mod traffic_shaping;
