+++ apollo-router/src/plugins/telemetry/apollo.rs
@@ -1,5 +1,5 @@
 //! Configuration for apollo telemetry.
-use crate::graphql::serde_utils::deserialize_header_name;
+use crate::graphql::plugin::utils::serde::deserialize_header_name;
 use http::header::HeaderName;
 use schemars::JsonSchema;
 use serde::Deserialize;
